+++ frontend/src/App.tsx
@@ -1,15 +1,10 @@
-import { Loading } from './components/Animation';
+import Loading from './components/Loading/Loading';
 import GlobalStyle from './styles/global';
 
 function App() {
   return (
     <>
       <GlobalStyle />
-      <p>IBM 마리아주 1 2 3 4 5 6 7 8 9 0</p>
-      <button>버튼</button>
-      <h1>Mariage</h1>
-      <h2>Mariage</h2>
-      <h3>Mariage</h3>
       <Loading />
     </>
   );
