--- frontend/src/components/Animation/Loading.tsx
@@ -1,14 +1,0 @@
-import Lottie from 'lottie-react';
-
-import loading from '../../assets/lottie/loading.json';
-
-const style = {
-  width: '50%',
-  maxHeight: '5rem',
-};
-
-function Loading() {
-  return <Lottie animationData={loading} style={style} />;
-}
-
-export default Loading;
