+++ frontend/src/components/Animation/index.ts
@@ -1,1 +1,1 @@
-export { default as Loading } from './Loading';
+export { default as Wine } from './Wine';
