+++ frontend/src/components/Loading/Loading.styled.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const Wrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export { Wrapper };
