+++ frontend/src/components/Loading/Loading.tsx
@@ -1,0 +1,13 @@
+import { Wine } from '../Animation';
+
+import * as S from './Loading.styled';
+
+function Loading() {
+  return (
+    <S.Wrapper>
+      <Wine />
+    </S.Wrapper>
+  );
+}
+
+export default Loading;
